tablero-nacho/requirements.txt
python-chess
flask
google-genai
python-dotenv
